pyproject.toml
[tool.poetry]
name = "sell-api"
version = "0.1.0"
description = ""
authors = ["Daniel Ale <danielale9291@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^5.0.6"
djangorestframework = "^3.15.1"
markdown = "^3.6"
django-filter = "^24.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
